// shared/__generated__/yoldi-api.ts
export interface LoginDto {
  email: string;
  password: string;
}

export interface ProfileDto {
  cover: ImageDto | null;
  description: string | null;
  email: string;
  image: ImageDto | null;
  name: string;
  slug: string;
}

export interface SignUpDto {
  email: string;
  name: string;
  password: string;
}